Handle dw as special instruction. Supports labels.
Downside is that it no longer handles list, but this is likely
only used for vectors anyway.

Theres a special case in the code, since the label rewrite typically
assumes an absolute addressed instruction. This means the address is at
one byte beyond the start of the instruction.

Theres no "opcode" byte for a dw, so it needs to be the same as
the instruction start rather than +1.